+++ main/src/ColorDetection.h
@@ -11,17 +11,17 @@
 
 extern Adafruit_TCS34725 tcs;
 
-uint8_t expectedRed[3] = {137, 68, 47};
-uint8_t expectedGreen[3] = {73, 115, 55};
-uint8_t expectedBlue[3] = {62, 82, 103};
+unsigned short expectedRed[3] = {137, 68, 47};
+unsigned short expectedGreen[3] = {73, 115, 55};
+unsigned short expectedBlue[3] = {62, 82, 103};
 
 #define AMOUNT_OF_COLORS 3
-uint8_t* expectedColors[AMOUNT_OF_COLORS] = {expectedRed, expectedGreen, expectedBlue};
-uint8_t colorCodes[AMOUNT_OF_COLORS] = {RED, GREEN, BLUE};
+unsigned short* expectedColors[AMOUNT_OF_COLORS] = {expectedRed, expectedGreen, expectedBlue};
+unsigned short colorCodes[AMOUNT_OF_COLORS] = {RED, GREEN, BLUE};
 
 #define DIFF_THRESHOLD 10
 
-void getColorCode(uint8_t *colorCode);
-void printColorCode(uint8_t colorCode);
+void getColorCode(unsigned short *colorCode);
+void printColorCode(unsigned short colorCode);
 
 #endif
